[ENH] refactored classification deep test to do something
