+++ aeon/classification/deep_learning/tests/test_deep_classifier_base.py
@@ -1,0 +1,71 @@
+# -*- coding: utf-8 -*-
+"""Unit tests for classifiers deep learning base class functionality."""
+
+import os
+
+import pytest
+
+from aeon.classification.deep_learning.base import BaseDeepClassifier
+from aeon.utils.validation._dependencies import _check_soft_dependencies
+
+__author__ = ["achieveordie", "hadifawaz1999"]
+
+
+class _DummyDeepClassifier(BaseDeepClassifier):
+    """Dummy Deep Classifier for testing empty base deep class save utilities."""
+
+    def __init__(self):
+        super(_DummyDeepClassifier, self).__init__()
+
+    def build_model(self, input_shape, n_classes, **kwargs):
+        import tensorflow as tf
+
+        input_layer = tf.keras.layers.Input(input_shape)
+        output_layer = tf.keras.layers.Dense(units=n_classes)(input_layer)
+
+        model = tf.keras.models.Model(inputs=input_layer, outputs=output_layer)
+
+        model.compile(loss="categorical_crossentropy")
+
+        return model
+
+    def _fit(self, X, y):
+        X = X.transpose(0, 2, 1)
+
+        self.input_shape_ = X.shape[1:]
+        self.model_ = self.build_model(self.input_shape_)
+
+        self.history = self.model_.fit(
+            X,
+            y,
+            batch_size=16,
+            epochs=2,
+        )
+
+        return self
+
+
+@pytest.mark.skipif(
+    not _check_soft_dependencies("tensorflow", severity="none"),
+    reason="skip test if required soft dependency not available",
+)
+def test_dummy_deep_classifier():
+    import numpy as np
+
+    # create a dummy deep classifier
+    dummy_deep_clf = _DummyDeepClassifier()
+
+    # test fit function on random data
+    dummy_deep_clf.fit(
+        X=np.random.normal(size=(10, 100)), y=np.random.choice([0, 1], size=(10,))
+    )
+
+    # test save last model to file than delete it
+
+    dummy_deep_clf.save_last_model_to_file()
+
+    os.remove("./last_model.hdf5")
+
+    # test summary of model
+
+    dummy_deep_clf.summary()
